aarch64 ABI 数据标签改 Darwin 本地 L 前缀；TLS 流终态原子转换显式化

+++ extlibs/xhttp/single/xhttp.h
@@ -176000,7 +176000,7 @@ static void __xrtTlsStreamFail(xtlsstream* pStream)
 	);
 	xrtAtomic32Store(
 		&pStream->TerminalResult,
-		XNET_RESULT_ERROR,
+		(uint32)XNET_RESULT_ERROR,
 		XMEMORY_RELEASE
 	);
 	xrtAtomic32Store(
@@ -176332,7 +176332,7 @@ static void __xrtTlsStreamProgressClose(xtlsstream* pStream)
 			);
 			xrtAtomic32Store(
 				&pStream->TerminalResult,
-				XNET_RESULT_ERROR,
+				(uint32)XNET_RESULT_ERROR,
 				XMEMORY_RELEASE
 			);
 			xrtAtomic32Store(
